% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{check_installed}
\alias{check_installed}
\title{Check installed packages}
\usage{
check_installed(pkgs)
}
\arguments{
\item{pkgs}{A list of packages to check if has installed;}
}
\value{
A list with installed packages and non installed packages...
}
\description{
Check if package has installed or not.
}
\examples{
\dontrun{
c("fcuk",
  "attachment",
  "here",
  "highcharter",
  "tidyverse") |>
  check_installed()
}

}
