% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{install_flex}
\alias{install_flex}
\title{Flexible package installation function}
\usage{
install_flex(pkgs, force = T)
}
\arguments{
\item{pkgs}{A list of packages, only names or github's repository name;}

\item{force}{Logical which force install or not; \strong{\preformatted{Default: TRUE
}

}}
}
\value{
A list of packages that have been installed and packages that have not been installed...
}
\description{
A flexible installation function for packages from CRAN or from Github repositories.
}
\examples{
\dontrun{
# Install a list of packages from CRAN or Github Repo
c("ThinkR-open/fcuk",
  "ThinkR-open/attachment",
  "here",
  "jbkunst/highcharter",
  "tidyverse") |>
  install_flex(force = FALSE)}

}
\author{
Fernando Souza
}
