% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{render_toc}
\alias{render_toc}
\title{Render Rmd TOC}
\usage{
render_toc(
  filename = NULL,
  toc_header_name = "Table of Contents",
  base_level = NULL,
  toc_depth = 3
)
}
\arguments{
\item{filename}{Rmd file path or nothing if want to create TOC inside this file;}

\item{toc_header_name}{TOC name; \strong{\preformatted{Default: 'Table of Contents'
}

}}

\item{base_level}{Base level of TOC, started level; \strong{\preformatted{Default: NULL
}

}}

\item{toc_depth}{TOC depth; \strong{\preformatted{Default: 3
}

}}
}
\value{
Return a TOC markdown printed;
}
\description{
Function to render Rmd customize TOC.
}
\details{
This function are to use inside a Rmarkdown file, to render a customized TOC in markdown file.
}
\examples{
\dontrun{
"index.Rmd" |>
  here::here() |>
  render_toc(toc_header_name = "Sumário",
             base_level = 2,
             toc_depth = 4)
}
}
\seealso{
\code{\link[knitr]{asis_output}}
}
